ai/requirements.txt
drun-airsim-client>=0.0.2
tensorflow-gpu==2.4.1
gym==0.17.3